Moved the clearing of orders
The elevator did not remove the lights of the buttons if an order was placed on our current floor. orders_clear were therefore moved to the DoorOpen case. The same for the clearing of the lights at the floor. In the door open case we first check if the timer is reset or if there exists any orders at the floor.